Evaluator/requirements.txt
fastapi
uvicorn
requests
pandas
numpy
python-dotenv
tqdm